fix!: incorrect slice axis after detach
Replace TranslateWithinPlane and TranslateAlongNormal with one TranslationController.
In reality, TranslationController is just TranslateWithinPlane renamed
